final_project/requirements.txt
tk
requests
pandas
mplfinance
matplotlib
Pillow